api/requirements.txt
Flask==2.0.1
Flask-JWT-Extended==4.4.2
Flask-Pydantic==0.9.0
Flask-SQLAlchemy==2.5.1
#google-cloud-bigquery==3.2.0
Jinja2==3.1.2
numpy==1.21.6
pandas==1.3.5
passlib==1.7.4
psycopg2-binary==2.9.1
pydantic==1.9.1
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
SQLAlchemy==1.4.25
